app/requirements.txt
pandas==1.2.4
pymongo==3.11.4
numpy==1.20.3
fastapi==0.65.2
python-multipart==0.0.5
openpyxl==3.0.7
PyYAML==5.4.1
uvicorn==0.14.0
requests==2.25.1
pytictoc==1.5.2
scikit-learn==0.24.2